Provide a wide QWERTY mapping
This way, the OS can map the layout to for instance Swedish

Standard layouts are actually _seven_ columns wide at least on the right half -	and because Miryoku is only five columns, it would make sense to be the reason Miryoku swaps semicolon with quote (apostrophe) which otherwise would not fit on the base layer. As can be seen in [keymap_extras/keymap_swedish.h](https://github.com/qmk/qmk_firmware/blob/master/quantum/keymap_extras/keymap_swedish.h), the OS would expect these to be in their normal places, in order to be _mapped_ to Å, Ä and Ö.

This would probably cause the keycodes also from the SYM layer to be remapped by the OS, so may require some further fiddling and certainly visualization. See the mapping section of [Miryoku Customisation Documentation](https://github.com/manna-harbour/miryoku/discussions/85)